+++ packages/core/examples/stackEvents/index.ts
@@ -67,38 +67,35 @@ let eventNumber = 1;
 
 const { IMAGE_RENDERED, CAMERA_MODIFIED, STACK_NEW_IMAGE } = Enums.Events;
 
-element.addEventListener(
-  IMAGE_RENDERED,
-  (evt: Types.EventTypes.ImageRenderedEvent) => {
-    updateLastEvents(eventNumber, IMAGE_RENDERED, JSON.stringify(evt.detail));
-    eventNumber++;
-  }
-);
+element.addEventListener(IMAGE_RENDERED, ((
+  evt: Types.EventTypes.ImageRenderedEvent
+) => {
+  updateLastEvents(eventNumber, IMAGE_RENDERED, JSON.stringify(evt.detail));
+  eventNumber++;
+}) as EventListener);
 
-element.addEventListener(
-  CAMERA_MODIFIED,
-  (evt: Types.EventTypes.CameraModifiedEvent) => {
-    updateLastEvents(eventNumber, CAMERA_MODIFIED, JSON.stringify(evt.detail));
-    eventNumber++;
-  }
-);
+element.addEventListener(CAMERA_MODIFIED, ((
+  evt: Types.EventTypes.CameraModifiedEvent
+) => {
+  updateLastEvents(eventNumber, CAMERA_MODIFIED, JSON.stringify(evt.detail));
+  eventNumber++;
+}) as EventListener);
 
-element.addEventListener(
-  STACK_NEW_IMAGE,
-  (evt: Types.EventTypes.StackNewImageEvent) => {
-    // Remove the image since then we serialize a bunch of pixelData to the screen.
-    const { imageId, renderingEngineId, viewportId } = evt.detail;
-    const detail = {
-      imageId,
-      renderingEngineId,
-      viewportId,
-      image: 'cornerstoneImageObject',
-    };
+element.addEventListener(STACK_NEW_IMAGE, ((
+  evt: Types.EventTypes.StackNewImageEvent
+) => {
+  // Remove the image since then we serialize a bunch of pixelData to the screen.
+  const { imageId, renderingEngineId, viewportId } = evt.detail;
+  const detail = {
+    imageId,
+    renderingEngineId,
+    viewportId,
+    image: 'cornerstoneImageObject',
+  };
 
-    updateLastEvents(eventNumber, STACK_NEW_IMAGE, JSON.stringify(detail));
-    eventNumber++;
-  }
-);
+  updateLastEvents(eventNumber, STACK_NEW_IMAGE, JSON.stringify(detail));
+  eventNumber++;
+}) as EventListener);
 
 addButtonToToolbar({
   title: 'Set VOI Range',