+++ packages/tools/src/eventDispatchers/keyboardToolEventDispatcher.ts
@@ -7,8 +7,8 @@ import { keyDown, keyUp } from './keyboardEventHandlers';
  * @param element - The HTML element to attach the event listeners to.
  */
 const enable = function (element: HTMLDivElement) {
-  element.addEventListener(Events.KEY_DOWN, keyDown);
-  element.addEventListener(Events.KEY_UP, keyUp);
+  element.addEventListener(Events.KEY_DOWN, keyDown as EventListener);
+  element.addEventListener(Events.KEY_UP, keyUp as EventListener);
 };
 
 /**
@@ -16,8 +16,8 @@ const enable = function (element: HTMLDivElement) {
  * @param element - The HTML element to attach the event listeners to.
  */
 const disable = function (element: HTMLDivElement) {
-  element.removeEventListener(Events.KEY_DOWN, keyDown);
-  element.removeEventListener(Events.KEY_UP, keyUp);
+  element.removeEventListener(Events.KEY_DOWN, keyDown as EventListener);
+  element.removeEventListener(Events.KEY_UP, keyUp as EventListener);
 };
 
 const keyboardToolEventDispatcher = {