+++ packages/tools/src/utilities/getModalityUnit.ts
@@ -1,0 +1,11 @@
+function getModalityUnit(modality: string, isPreScaled: boolean): string {
+  if (modality === 'CT') {
+    return 'HU';
+  } else if (modality === 'PT' && isPreScaled === true) {
+    return 'SUV';
+  } else {
+    return '';
+  }
+}
+
+export { getModalityUnit };